Extract blob files
