+++ src/app/page.module.scss
@@ -68,7 +68,7 @@
 }
 
 .blobs {
-  grid-area: "content";
+  grid-area: content;
   height: 400px;
   width: 400px;
   position: relative;
